requirements.txt
flask
yfinance
finnhub-python
mysql-connector-python
schedule
gunicorn